+++ EShop.Presentation/ProductDto.cs
@@ -1,0 +1,8 @@
+﻿namespace EShop.Presentation
+{
+    public record ProductDto
+    {
+        public decimal Price { get; set; }
+        public string Name { get; set; }
+    }
+}
